+++ api/serializers.py
@@ -30,20 +30,6 @@ from threepio import logger
 
 
 # Serializers
-class PaginatedApplicationSerializer(pagination.PaginationSerializer):
-    """
-    Serializes page objects of Instance querysets.
-    """
-
-    def __init__(self, *args, **kwargs):
-        user = get_context_user(self, kwargs)
-        self.request_user = user
-        super(PaginatedApplicationSerializer, self).__init__(*args, **kwargs)
-
-    class Meta:
-        object_serializer_class = ApplicationSerializer
-
-
 class ApplicationBookmarkSerializer(serializers.ModelSerializer):
     """
     """
