+++ api/serializers/__init__.py
@@ -12,3 +12,4 @@ from .cleaned_identity_serializer import CleanedIdentitySerializer
 from .boot_script_serializer import BootScriptSerializer
 from .application_threshold_serializer import ApplicationThresholdSerializer
 from .application_serializer import ApplicationSerializer
+from .paginated_application_serializer import PaginatedApplicationSerializer
